refactor: refactor event endpoint schemas to its own file + add event descriptions
